+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ##Change Log
 
+#### Version 0.1.7 January 2016
+- **(C)** Extended UIkit plus features 
+- **(C)** Updated UIkit to version 2.24.3
+
 #### Version 0.1.6 December 2015
+- **(A)** Added UIkit Sticky support
 - **(B)** Removed getDocument call from onAfterInitialise
 
 #### Version 0.1.5 December 2015